# pkg/docs/swagger.yaml
basePath: /dragonfly
definitions:
  agent.AgentInfo:
    properties:
      agent_health:
        type: string
      agent_state:
        type: string
      agent_type:
        type: string
      agent_unhealthy_resp_cnt:
        type: integer
      csp_type:
        type: string
      mcis_id:
        type: string
      ns_id:
        type: string
      public_ip:
        type: string
      vm_id:
        type: string
    type: object
  agent.JSONResult:
    type: object
...
  rest.AgentMetaDataListType:
    properties:
      id(ns_id/mcis_id/vm_id/csp_type):
        $ref: '#/definitions/agent.AgentInfo'
    type: object
  rest.AgentType:
    properties:
      cspType:
        type: string
      mcis_id:
        type: string
      ns_id:
        type: string
      port:
        type: string
      public_ip:
        type: string
      ssh_key:
        type: string
      user_name:
        type: string
      vm_id:
        type: string
    type: object
  rest.SimpleMsg:
    properties:
      message:
        example: Any message
        type: string
    type: object
...
  types.AlertEventHandler:
    properties:
      id:
        type: string
      name:
        type: string
      options:
        additionalProperties: true
        type: object
      type:
        type: string
    type: object
  types.AlertEventHandlerReq:
    properties:
      channel:
        type: string
      from:
        type: string
      host:
        description: Parameters for SMTP
        type: string
      name:
        type: string
      password:
        type: string
      port:
        type: integer
      to:
        items:
          type: string
        type: array
      type:
        type: string
      url:
        description: Parameters for Slack
        type: string
      username:
        type: string
      workspace:
        type: string
    type: object
...
host: localhost:9090
info:
  contact:
    email: contact-to-cloud-barista@googlegroups.com
    name: API Support
    url: http://cloud-barista.github.io
  description: CB-Dragonfly REST API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: CB-Dragonfly REST API
  version: "1.0"
paths:
  /agent:
    delete:
      consumes:
      - application/json
      description: 모니터링 에이전트 제거
      parameters:
      - description: Details for an Agent Remove object
        in: body
        name: agentInfo
        required: true
        schema:
          $ref: '#/definitions/rest.AgentType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.SimpleMsg'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.SimpleMsg'
      summary: Uninstall agent to vm
      tags:
      - '[Agent] Monitoring Agent'
    post:
      consumes:
      - application/json
      description: 모니터링 에이전트 설치
      parameters:
      - description: Details for an Agent Install object
        in: body
        name: agentInfo
        required: true
        schema:
          $ref: '#/definitions/rest.AgentType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.SimpleMsg'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.SimpleMsg'
      summary: Install agent to vm
      tags:
      - '[Agent] Monitoring Agent'
  /agent/metadata:
    get:
      consumes:
      - application/json
      description: 에이전트 메타데이터 조회
      parameters:
      - description: 네임스페이스 아이디
        enum:
        - test_ns
        in: query
        name: ns
        type: string
      - description: MCIS 아이디
        enum:
        - test_mcis
        in: query
        name: mcisId
        type: string
      - description: VM 아이디
        enum:
        - test_vm
        in: query
        name: vmId
        type: string
      - description: VM의 CSP 정보
        enum:
        - aws
        in: query
        name: cspType
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Different return structures by the given param
          schema:
            allOf:
            - $ref: '#/definitions/agent.JSONResult'
            - properties:
                '[DEFAULT]':
                  items:
                    $ref: '#/definitions/rest.AgentMetaDataListType'
                  type: array
                '[ID]':
                  $ref: '#/definitions/agent.AgentInfo'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.SimpleMsg'
      summary: List agent metadata
      tags:
      - '[Agent] Monitoring Agent'
  /alert/eventhandler:
    post:
      consumes:
      - application/json
      description: 알람 이벤트 핸들러 생성
      parameters:
      - description: Details for an EventHandler object
        in: body
        name: eventHandlerInfo
        required: true
        schema:
          $ref: '#/definitions/types.AlertEventHandlerReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.AlertEventHandler'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.SimpleMsg'
      summary: Create monitoring alert event-handler
      tags:
      - '[EventHandler] Alarm Event Handler management'
  /alert/eventhandler/type/{type}/event/{name}:
    delete:
      consumes:
      - application/json
      description: 알람 이벤트 핸들러 삭제
      parameters:
      - description: 이벤트 핸들러 유형
        in: path
        name: type
        required: true
        type: string
      - description: 이벤트 핸들러 이름
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.SimpleMsg'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.SimpleMsg'
      summary: Delete monitoring alert event-handler
      tags:
      - '[EventHandler] Alarm Event Handler management'
    get:
      consumes:
      - application/json
      description: 알람 이벤트 핸들러 조회
      parameters:
      - description: 이벤트 핸들러 유형
        in: path
        name: type
        required: true
        type: string
      - description: 이벤트 핸들러 이름
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.AlertEventHandler'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.SimpleMsg'
      summary: Get monitoring alert event-handler
      tags:
      - '[EventHandler] Alarm Event Handler management'
    put:
      consumes:
      - application/json
      description: 알람 이벤트 핸들러 수정
      parameters:
      - description: 이벤트 핸들러 유형
        in: path
        name: type
        required: true
        type: string
      - description: 이벤트 핸들러 이름
        in: path
        name: name
        required: true
        type: string
      - description: Details for an EventHandler (slack) object
        in: body
        name: eventHandlerInfo
        required: true
        schema:
          $ref: '#/definitions/types.AlertEventHandlerReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.AlertEventHandler'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.SimpleMsg'
      summary: Update monitoring alert event-handler
      tags:
      - '[EventHandler] Alarm Event Handler management'
  /alert/eventhandlers:
    get:
      consumes:
      - application/json
      description: 알람 이벤트 핸들러 목록 조회
      parameters:
      - description: 이벤트 핸들러 유형
        enum:
        - slack
        - smtp
        in: query
        name: eventType
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.AlertEventHandler'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.SimpleMsg'
      summary: List monitoring alert event-handler
      tags:
      - '[EventHandler] Alarm Event Handler management'
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"